LPS-150272 Move package, not only for upgrading logic
