+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/util/GradleBuildFile.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.source.formatter.upgrade;
+package com.liferay.source.formatter.util;
 
 import com.liferay.petra.string.StringUtil;
 import com.liferay.portal.kernel.log.Log;
