+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/util/GradleDependency.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.source.formatter.upgrade;
+package com.liferay.source.formatter.util;
 
 import java.text.MessageFormat;
